// collect/company/ems/admin/src/pages/relation-analysis/ra-global-parameter-manage/ra-gpm-point-select/ra-gpm-point-select.api.ts
export interface RA_GPM_IPointSelectInfo {
  deviceId: string;
  deviceName: string;
  pointId: string;
  pointName: string;
}

export interface RA_GPM_IDeviceInfo {
  deviceId: number;
  deviceName: string;
  id: number;
  paramName: string;
  pointName: string;
  pointNumber: number;
}

export interface RA_GPM_IPointInfo {}